+++ src/main/java/domain/GameResults.java
@@ -1,6 +1,7 @@
 package domain;
 
 import domain.constant.GameResult;
+import domain.participant.Player;
 
 import java.util.List;
 import java.util.Map;
