--- src/main/java/domain/Participant.java
@@ -1,11 +1,0 @@
-package domain;
-
-public interface Participant {
-    boolean isDrawable();
-
-    void draw(final Deck deck);
-
-    default boolean isDealer() {
-        return this instanceof Dealer;
-    }
-}
