+++ src/main/java/domain/participant/Dealer.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.participant;
 
+import domain.*;
 import domain.constant.GameResult;
 
 import java.util.ArrayList;
@@ -7,13 +8,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class Dealer implements Participant {
+public class Dealer extends Participant {
     private static final int MAX_DEALER_HAND_VALUE = 16;
 
-    private final Hand hand;
-
     Dealer(final Hand hand) {
-        this.hand = hand;
+        super(hand);
     }
 
     public static Dealer init() {
@@ -39,7 +38,7 @@ public class Dealer implements Participant {
         return new GameResults(dealerGameResults, playerGameResults);
     }
 
-    private void match(Player player, List<GameResult> dealerGameResults, Map<Player, GameResult> playerGameResults) {
+    private void match(final Player player, final List<GameResult> dealerGameResults, final Map<Player, GameResult> playerGameResults) {
         if (isDealerWin(player)) {
             dealerWin(player, dealerGameResults, playerGameResults);
             return;
@@ -48,15 +47,15 @@ public class Dealer implements Participant {
     }
 
     private boolean isDealerWin(final Player player) {
-        return hand.isBlackJack() || (hand.getCardsNumberSum() > player.getPlayerHandStatus().cardNumberSum());
+        return hand.isBlackJack() || (hand.getCardsNumberSum() > player.getHandStatus().cardNumberSum());
     }
 
-    private void dealerWin(Player player, List<GameResult> dealerGameResults, Map<Player, GameResult> playerGameResults) {
+    private void dealerWin(final Player player, final List<GameResult> dealerGameResults, final Map<Player, GameResult> playerGameResults) {
         dealerGameResults.add(GameResult.WIN);
         playerGameResults.put(player, GameResult.LOSE);
     }
 
-    private void dealerLose(Player player, List<GameResult> dealerGameResults, Map<Player, GameResult> playerGameResults) {
+    private void dealerLose(final Player player, final List<GameResult> dealerGameResults, final Map<Player, GameResult> playerGameResults) {
         dealerGameResults.add(GameResult.LOSE);
         playerGameResults.put(player, GameResult.WIN);
     }
