+++ src/main/java/domain/participant/Hand.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.participant;
+
+import domain.PlayingCard;
 
 import java.util.ArrayList;
 import java.util.List;
