+++ src/main/java/domain/participant/HandStatus.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.participant;
+
+import domain.PlayingCard;
 
 import java.util.List;
