+++ src/main/java/view/InputView.java
@@ -11,7 +11,9 @@ public class InputView {
     private static final Scanner scanner = new Scanner(System.in);
 
     public static List<String> inputPlayerNames() {
+        System.out.println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으롭 분리)");
         String inputString = scanner.nextLine();
+        System.out.println();
         return Arrays.stream(inputString.split(",")).toList();
     }
