+++ src/main/java/view/OutputView.java
@@ -5,4 +5,6 @@ public class OutputView {
     public static void printErrorMessage(String message) {
         System.out.println(message);
     }
+
+//    public static void
 }
