+++ src/test/java/domain/participant/ParticipantTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
