+++ src/test/java/domain/participant/PlayerNameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
