+++ src/test/java/domain/participant/PlayerNamesTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.participant;
 
+import domain.participant.PlayerNames;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
